Refactored floating point emulation
Refactored for better 64-bit value loading and storing. I also added missing opcodes.

+++ main.cpp
@@ -68,7 +68,7 @@ uint32_t ppc_effective_address;
 uint32_t ppc_real_address;
 uint32_t ppc_next_instruction_address; //Used for branching, setting up the NIA
 
-uint32_t return_value;
+uint64_t return_value;
 
 MemCtrlBase *mem_ctrl_instance = 0;
 HeathrowIC  *heathrow = 0;
